// src/api/types.gen.ts
export type PaginationCursor = {
    next?: string;
} | null;

/**
 * The behavior referenced by the policy, that is validated by the policy.
 */
export type BehaviorPolicyBehavior = {
    /**
     * The  arguments that the referenced validator type requires.
     */
    arguments?: {
        [key: string]: unknown;
    };
    /**
     * The unique identifier of a pre-defined behavior.
     */
    id: string;
};
// ...
export type BehaviorPolicyDeserialization = {
    publish?: BehaviorPolicyDeserializer;
    will?: BehaviorPolicyDeserializer;
};

/**
 * The matching rules the policy applies.
 */
export type BehaviorPolicyMatching = {
    /**
     * The regex pattern to match the client id against.
     */
    clientIdRegex: string;
};
// ...
export type BehaviorPolicyOnTransition = {
    'Connection.OnDisconnect'?: BehaviorPolicyOnEvent;
    'Event.OnAny'?: BehaviorPolicyOnEvent;
    'Mqtt.OnInboundConnect'?: BehaviorPolicyOnEvent;
    'Mqtt.OnInboundDisconnect'?: BehaviorPolicyOnEvent;
    'Mqtt.OnInboundPublish'?: BehaviorPolicyOnEvent;
    'Mqtt.OnInboundSubscribe'?: BehaviorPolicyOnEvent;
    /**
     * The exact state from which the transition happened. Alternatively a state filter can be used.
     */
    fromState: string;
    /**
     * The exact state to which the transition happened. Alternatively a state filter can be used.
     */
    toState: string;
};

/**
 * A policy which is used to validate and execute certain actions based on the validation result.
 */
export type BehaviorPolicy = {
    behavior: BehaviorPolicyBehavior;
    /**
     * The formatted UTC timestamp indicating when the policy was created.
     */
    readonly createdAt?: string;
    deserialization?: BehaviorPolicyDeserialization;
    /**
     * The unique identifier of the policy.
     */
    id: string;
    /**
     * The formatted UTC timestamp indicating when the policy was updated the last time.
     */
    readonly lastUpdatedAt?: string;
    matching: BehaviorPolicyMatching;
    onTransitions?: Array<BehaviorPolicyOnTransition>;
};

/**
 * A listing of behavior policies.
 */
export type BehaviorPolicyList = {
    _links?: PaginationCursor;
    /**
     * List of result items that are returned by this endpoint
     */
    items?: Array<BehaviorPolicy>;
};
// ...
export type DataPolicyMatching = {
    /**
     * The topic filter for which the policy is matched.
     */
    topicFilter: string;
};

/**
 * One or more operations the outcome of the validation triggers.  When this field is empty, the outcome of the policy validation does not trigger any operations.
 */
export type DataPolicyAction = {
    /**
     * The pipeline to execute, when this action is triggered. The operations in the pipeline are executed in-order.
     */
    pipeline?: Array<PolicyOperation>;
};
// ...
export type DataPolicyValidation = {
    /**
     * The validators of the policy.
     */
    validators?: Array<DataPolicyValidator>;
};

/**
 * A data policy which is used to validate and execute certain actions based on the validation result.
 */
export type DataPolicy = {
    /**
     * The formatted UTC timestamp indicating when the policy was created.
     */
    readonly createdAt?: string;
    /**
     * The unique identifier of the policy.
     */
    id: string;
    /**
     * The formatted UTC timestamp indicating when the policy was updated the last time.
     */
    readonly lastUpdatedAt?: string;
    matching: DataPolicyMatching;
    onFailure?: DataPolicyAction;
    onSuccess?: DataPolicyAction;
    validation?: DataPolicyValidation;
};

/**
 * A listing of data policies.
 */
export type DataPolicyList = {
    _links?: PaginationCursor;
    /**
     * List of result items that are returned by this endpoint
     */
    items?: Array<DataPolicy>;
};
// ...
export type PolicySchema = {
    /**
     * The schema type dependent arguments.
     */
    arguments?: {
        [key: string]: string;
    };
    /**
     * The formatted UTC timestamp when the schema was created.
     */
    readonly createdAt?: string;
    /**
     * The unique identifier of the schema.
     */
    id: string;
    /**
     * The base64 encoded schema definition.
     */
    schemaDefinition: string;
    /**
     * The type of the schema.
     */
    type: string;
    /**
     * The version of the schema.
     */
    readonly version?: number;
};

/**
 * A listing of schemas.
 */
export type SchemaList = {
    _links?: PaginationCursor;
    /**
     * List of result items that are returned by this endpoint
     */
    items?: Array<PolicySchema>;
};

export type Script = {
    /**
     * The formatted UTC timestamp when the script was created.
     */
    readonly createdAt?: string;
    /**
     * A string of free-form text describing the function.
     */
    description?: string;
    /**
     * The type of the function.
     */
    functionType: 'TRANSFORMATION';
    /**
     * The unique identifier of the script.
     */
    id: string;
    /**
     * The base64 encoded function source code.
     */
    source: string;
    /**
     * The version of the script.
     */
    readonly version?: number;
};

/**
 * A listing of scripts.
 */
export type ScriptList = {
    _links?: PaginationCursor;
    /**
     * List of result items that are returned by this endpoint
     */
    items?: Array<Script>;
};